Fix decimal in DB and Logic

+++ Application/DataTransferObjects/Lsn50V2Lifecycle/CreateLsn50V2LifecycleDto.cs
@@ -3,6 +3,6 @@
 public class CreateLsn50V2LifecycleDto
 {
     public Guid SensorId { get; set; }
-    public double BatteryVoltage { get; set; }
+    public decimal BatteryVoltage { get; set; }
     public int BatteryLevel { get; set; }
 }
